refactor(engine,scoring): split native_executor.go and scoring/engine.go into single-responsibility files
native_executor.go (1263→353 lines) split into:
- executor_builders.go: message/prompt/config builders, tool dispatch
- executor_sandbox.go: sandbox lifecycle, policy, configuration
- executor_staging.go: workspace input staging, fixture file handling
- executor_retry.go: provider retry logic
- executor_util.go: clone helpers, data utilities

scoring/engine.go (1442→196 lines) split into:
- engine_evidence.go: evidence extraction from events
- engine_validators.go: validator evaluation
- engine_metrics.go: metric collection
- engine_dimensions.go: dimension scoring and normalization
- engine_pricing.go: model cost calculation
- engine_resolution.go: evidence source resolution
- engine_util.go: payload parsing, type conversion utilities

No behavior changes. All existing tests pass without modification.

Closes #225
